Create news_to_gcs.py
Here is the dag that "works on my PC". 
It gets the list of URLs form the 'rss_feed.csv' file. Parses it in memory. 
For each URLs it founds it gets the corresponding RSS feeds and saves it in raw format. 
the files in raw format are then pre-processed in memory and then written back to the bucket as csv.